ICMPv6 MAG Driver
The ICMP MAG driver adds support for mobile node attach events based
on router solicitations. This allows the MAG to operate on any access
technology, albeit with some limitations in the attachment state.

Future Work:
Use a timer to periodicaly check (before the binding lifetime expires) the
presence of the mobile node.